+++ Backend/routes.js
@@ -1,75 +1,82 @@
-const express = require('express');
-const userDetails = require('./models/models'); 
-const { get } = require('mongoose');
+const express = require("express");
+const userDetails = require("./models/models");
+const { get } = require("mongoose");
 
-const getRouter = express.Router(); 
+const getRouter = express.Router();
 
-getRouter.get('/GET', async (req, res) => {
-    try {
-        const details = await userDetails.find();
-        
+getRouter.get("/GET", async (req, res) => {
+  try {
+    const details = await userDetails.find();
 
-        console.log('Retrieved details:', details);
+    console.log("Retrieved details:", details);
 
-
-        if (details.length === 0) {
-            return res.status(404).json({ message: 'No details found' });
-        }
-
-
-        res.status(200).json(details);
-    } catch (err) {
-        console.error(err, 'GET error');
-        res.status(500).json({ error: 'Internal Server Error' });
+    if (details.length === 0) {
+      return res.status(404).json({ message: "No details found" });
     }
+
+    res.status(200).json(details);
+  } catch (err) {
+    console.error(err, "GET error");
+    res.status(500).json({ error: "Internal Server Error" });
+  }
 });
 
-getRouter.post('/POST', async (req, res)=>{
-    try{
-        const {name, country, age, content} = req.body;
-        console.log(req.body);
-        const newUser = new userDetails({
-            name, 
-            country, 
-            age, 
-            content
-        });
-        await newUser.save();
-        res.status(200).json({"Successfully added new user": newUser});
-
-    }catch(err){
-        console.error(err, "Post error")
-    }
-})
-getRouter.patch('/PATCH/:userid', async (req,res)=>{
-    const {userid} =  req.params;
-    const updates = req.body;
-
-    try{
-        const updateuser = await userDetails.findOneAndUpdate(
-            {userid: userid},
-            {$set: updates},
-            {new: true}
-        );
-        res.status(200).json({message:"successfully updated user", updateuser });
-
-    }catch(err){
-        console.error(err, "patch error")
-    }
+getRouter.post("/POST", async (req, res) => {
+  try {
+    const { name, country, age, content } = req.body;
+    console.log(req.body);
+    const newUser = new userDetails({
+      name,
+      country,
+      age,
+      content,
+    });
+    await newUser.save();
+    res.status(200).json({ "Successfully added new user": newUser });
+  } catch (err) {
+    console.error(err, "Post error");
+  }
 });
 
-getRouter.delete('/DELETE/:userid', async (req,res)=>{
-    const {userid} =  req.params;
+getRouter.patch("/PATCH/:id", async (req, res) => {
+  const {id} = req.params;
+  console.log(id)
+  const updates = req.body;
 
-    try{
-        const deleteuser = await userDetails.findOneAndDelete(
-            {userid: userid},
-        );
-        res.status(200).json({message:"successfully Deleted user", deleteuser });
+  try {
+    const updatedUser = await userDetails.findOneAndUpdate(
+      { _id : id }, 
+      { $set: updates },
+      { new: true }
+    );
 
-    }catch(err){
-        console.error(err, "delete error")
+    if (!updatedUser) {
+      return res.status(404).json({ message: "User not found" });
     }
+
+    res.status(200).json({ message: "Successfully updated user", updatedUser });
+  } catch (err) {
+    console.error(err, "patch error");
+    res.status(500).json({ error: "Internal Server Error" });
+  }
+});
+
+
+getRouter.delete("/DELETE/:id", async (req, res) => {
+  const { id: _id } = req.params;
+  // console.log(_id, "_id");
+  try {
+    const deleteuser = await userDetails.findOneAndDelete({ _id }); 
+    if (!deleteuser) {
+      return res.status(404).json({ message: "User not found" });
+    } 
+    
+    res.status(200).json({ message: "successfully Deleted user", deleteuser });
+    console.log(deleteuser)
+    
+  } catch (err) {
+    console.error(err, "delete error");
+  }
 });
 
 module.exports = { getRouter };
