+++ Client/data.json
@@ -1,0 +1,8 @@
+[
+    {
+        "name": "Mohammed Rafeeq Shariff",
+        "squad": "Squad_51",
+        "email": "mohammed.rafeeq@kalvium.community",
+        "github": "mohammedrafeeqshariff"
+    }
+]
