+++ Client/src/App.css
@@ -15,6 +15,5 @@
 
 body{
     background-color: black;
-    color: white;
-    user-select: none;
+    color: white;    
 }
