+++ Client/src/components/UploadData/Upload.css
@@ -31,7 +31,9 @@
     padding: 2%;
     width: 30%;
     margin: auto;
+    cursor: pointer;
 }
+
 
 input[type="number"]::-webkit-inner-spin-button,
 input[type="number"]::-webkit-outer-spin-button {
